Remove LoanBorrowDetailsScene
-Remove scene
-Migrate controls to LoanDetailsScene
